requirements.txt
scrapy
scrapy-fake-useragent
scrapy_proxies
scrapy-proxies-tool
requests
scrapy-rotating-proxies
scrapy-html-storage
beautifulsoup4
python-dotenv
flask
